refactor: adapt changes in the tree for the refactored button
